Rename framework to WireLinkPreview
