+++ Resources/Configurations/WireLinkPreview-ios.xcconfig
@@ -1,0 +1,2 @@
+#include "WireLinkPreview.xcconfig"
+#include "zmc-config/ios.xcconfig"
