--- Resources/Configurations/ZMCLinkPreview-ios.xcconfig
@@ -1,2 +1,0 @@
-#include "ZMCLinkPreview.xcconfig"
-#include "zmc-config/ios.xcconfig"
