+++ Scripts/config.txt
@@ -1,4 +1,4 @@
-framework=ZMCLinkPreview
+framework=WireLinkPreview
 repository=wire-ios-link-preview
 release-name-format=The {BUZZWORD} {BUZZWORD} {ANIMAL}
 organization=wireapp
