+++ WireLinkPreview/ImageDownloader.swift
