+++ WireLinkPreview/LinkPreviewDetector.swift
