+++ WireLinkPreview/LinkPreviewTypes.swift
