+++ WireLinkPreview/MetaStreamContainer.swift
