+++ WireLinkPreview/OpenGraphDataTypes.swift
