+++ WireLinkPreview/PreviewDownloader.swift
