+++ WireLinkPreviewTests/ImageDownloaderTests.swift
@@ -18,7 +18,7 @@
 
 
 import XCTest
-@testable import ZMCLinkPreview
+@testable import WireLinkPreview
 
 class ImageDownloaderTests: XCTestCase {
