+++ WireLinkPreviewTests/PreviewBlackListTests.swift
@@ -18,30 +18,28 @@
 
 
 import XCTest
-@testable import ZMCLinkPreview
+@testable import WireLinkPreview
 
-class ZMCLinkPreviewTests: XCTestCase {
+class PreviewBlackListTests: XCTestCase {
+    
+    var sut: PreviewBlacklist!
     
     override func setUp() {
         super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
+        sut = PreviewBlacklist()
     }
     
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
-    
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-    
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measureBlock {
-            // Put the code you want to measure the time of here.
+    func testThatCorrectHostsAreAllBlacklisted() {
+        let hosts = ["soundcloud", "vimeo", "spotify", "youtube", "giphy", "youtu.be", "y2u.be"]
+        
+        for host in hosts {
+            assertThatHostIsBlacklisted(host)
         }
+    }
+    
+    func assertThatHostIsBlacklisted(_ host: String, line: UInt = #line) {
+        let url = URL(string: "www.\(host).com/example")!
+        XCTAssertTrue(sut.isBlacklisted(url), "\(host) was not blacklisted", line: line)
     }
     
 }
