+++ WireLinkPreviewTests/URLMocks.swift
@@ -16,7 +16,7 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-@testable import ZMCLinkPreview
+@testable import WireLinkPreview
 
 class MockURLSessionDataTask: URLSessionDataTaskType {
