+++ WireLinkPreviewTests/polygon_head.txt
